[PC-190] state의 모델을 도메인 모델로 변경
