--- core/domain/src/main/java/com/puzzle/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.puzzle.domain
-
-class MyClass {
-}
