+++ core/domain/src/main/java/com/puzzle/domain/model/value/ValueTalkCard.kt
@@ -1,0 +1,7 @@
+package com.puzzle.domain.model.value
+
+data class ValueTalkCard(
+    val label: String = "",
+    val title: String = "",
+    val content: String = "",
+)
